pyproject.toml
[tool.poetry]
name = "tutorial"
version = "0.1.0"
description = ""
authors = ["Michal Nowak <m.nowak@bankier.pl >"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^3.8.1"
drf-extensions = "^0.7.1"
psycopg2-binary = "^2.9.6"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"